+++ Test/Source/Matchers/ConstraintEqualMatcher.swift
@@ -9,12 +9,12 @@ import Hamcrest
 import UIKit
 
 
-private func hasMatchingEqualConstraint(_ view: UIView, attribute: NSLayoutConstraint.Attribute, constant: Float = 0) -> MatchResult {
+private func hasMatchingEqualConstraint(_ view: UIView, attribute: NSLayoutConstraint.Attribute, constant: CGFloat = 0) -> MatchResult {
 	return hasMatchingEqualConstraint(view, to: view.superview, attribute: attribute, constant: constant)
 }
 
 
-private func hasMatchingEqualConstraint(_ view: UIView, to toView: UIView?, attribute: NSLayoutConstraint.Attribute, constant: Float = 0) -> MatchResult {
+private func hasMatchingEqualConstraint(_ view: UIView, to toView: UIView?, attribute: NSLayoutConstraint.Attribute, constant: CGFloat = 0) -> MatchResult {
 
 	if let superView = findSuperView(view, toView) {
 		var secondItem = toView
@@ -29,7 +29,7 @@ private func hasMatchingEqualConstraint(_ view: UIView, to toView: UIView?, attr
 					constraint.firstAttribute == attribute &&
 					constraint.secondAttribute == attribute &&
 					constraint.relation == NSLayoutConstraint.Relation.equal &&
-					constraint.constant == CGFloat(constant) &&
+					constraint.constant == constant &&
 					constraint.isActive) {
 				return .match
 			}
@@ -55,7 +55,7 @@ private func hasMatchingEqualConstraint(_ view: UIView, secondView: UIView?, fir
 	return hasMatchingEqualConstraint(view, secondView: secondView, firstAttribute: firstAttribute, secondAttribute: secondAttribute, multiplier: 1.0)
 }
 
-private func hasMatchingEqualConstraint(_ view: UIView, secondView: UIView?, firstAttribute: NSLayoutConstraint.Attribute, secondAttribute: NSLayoutConstraint.Attribute, multiplier: Float) -> MatchResult {
+private func hasMatchingEqualConstraint(_ view: UIView, secondView: UIView?, firstAttribute: NSLayoutConstraint.Attribute, secondAttribute: NSLayoutConstraint.Attribute, multiplier: CGFloat) -> MatchResult {
 
 
 	if let superView = view.superview {
@@ -104,14 +104,14 @@ private func hasMatchingEqualConstraint(_ view: UIView, secondView: UIView?, fir
 }
 
 
-public func hasEqualConstraint<T: UIView>(_ attribute: NSLayoutConstraint.Attribute, withConstant constant: Float = 0) -> Matcher<T> {
+public func hasEqualConstraint<T: UIView>(_ attribute: NSLayoutConstraint.Attribute, withConstant constant: CGFloat = 0) -> Matcher<T> {
 	return Matcher("view has equal constraint: \(descriptionOfAttribute(attribute))") {
 		(value: T) -> MatchResult in
 		return hasMatchingEqualConstraint(value, attribute: attribute, constant: constant)
 	}
 }
 
-public func hasEqualConstraint<T: UIView>(_ attribute: NSLayoutConstraint.Attribute, with view: UIView, constant: Float = 0) -> Matcher<T> {
+public func hasEqualConstraint<T: UIView>(_ attribute: NSLayoutConstraint.Attribute, with view: UIView, constant: CGFloat = 0) -> Matcher<T> {
 	return Matcher("view has equal constraint: \(descriptionOfAttribute(attribute))") {
 		(value: T) -> MatchResult in
 		return hasMatchingEqualConstraint(value, to: view, attribute: attribute, constant: constant)
@@ -122,11 +122,11 @@ public func isVerticalCenter<T: UIView>() -> Matcher<T> {
 	return hasEqualConstraint(.centerY)
 }
 
-public func isVerticalCenter<T: UIView>(offset: Float) -> Matcher<T> {
+public func isVerticalCenter<T: UIView>(offset: CGFloat) -> Matcher<T> {
 	return hasEqualConstraint(.centerY, withConstant: offset)
 }
 
-public func isVerticalCenter<T: UIView>(with view: UIView, offset: Float = 0) -> Matcher<T> {
+public func isVerticalCenter<T: UIView>(with view: UIView, offset: CGFloat = 0) -> Matcher<T> {
 	return hasEqualConstraint(.centerY, with: view, constant: offset)
 }
 
@@ -134,11 +134,11 @@ public func isHorizontalCenter<T: UIView>() -> Matcher<T> {
 	return hasEqualConstraint(.centerX)
 }
 
-public func isHorizontalCenter<T: UIView>(offset: Float) -> Matcher<T> {
+public func isHorizontalCenter<T: UIView>(offset: CGFloat) -> Matcher<T> {
 	return hasEqualConstraint(.centerX, withConstant: offset)
 }
 
-public func isHorizontalCenter<T: UIView>(with view: UIView, offset: Float = 0) -> Matcher<T> {
+public func isHorizontalCenter<T: UIView>(with view: UIView, offset: CGFloat = 0) -> Matcher<T> {
 	return hasEqualConstraint(.centerX, with: view, constant: offset)
 }
 
@@ -164,7 +164,7 @@ public func hasSameWidthAs<T: UIView>(_ view: UIView) -> Matcher<T> {
 	return hasSameWidthAs(view, multiplier: 1.0)
 }
 
-public func hasSameWidthAs<T: UIView>(_ view: UIView, multiplier: Float) -> Matcher<T> {
+public func hasSameWidthAs<T: UIView>(_ view: UIView, multiplier: CGFloat) -> Matcher<T> {
 	return Matcher("view has same size") {
 		(value: T) -> MatchResult in
 		return hasMatchingEqualConstraint(value, secondView: view, firstAttribute: .width, secondAttribute: .width, multiplier: multiplier)
@@ -179,7 +179,7 @@ public func hasSameHeightAs<T: UIView>(_ view: UIView) -> Matcher<T> {
 }
 
 
-public func hasSameHeightAs<T: UIView>(_ view: UIView, multiplier: Float) -> Matcher<T> {
+public func hasSameHeightAs<T: UIView>(_ view: UIView, multiplier: CGFloat) -> Matcher<T> {
 	return Matcher("view has same size") {
 		(value: T) -> MatchResult in
 		return hasMatchingEqualConstraint(value, secondView: view, firstAttribute: .height, secondAttribute: .height, multiplier: multiplier)
